requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-engineio==0.8.6
python-socketio==1.0
PyYAML==3.11
requests==2.9.1
six==1.10.0
slacker==0.9.0
socketIO-client==0.6.5
websocket-client==0.35.0
Werkzeug==0.11.3
